refactor: split auth form components by protocol
